Show draggable control bar in replay view (#988)
* Fix links in changelog (#977)

* Update non breaking dependencies (#985)

- addons-linter
- eslint-plugin-mozilla
- eslint-plugin-react
- husky
- react-player
- webpack

Also needed to add a new dep:
- eslint-no-unsanitized

Looks like this is a dep of eslint-plugin-mozilla that isn't listed correctly in the
eslint-plugin-mozilla package. This package is for some reason bundled in with gecko
so I do not have the time to contribute a fix at this time. Hopefully I will find
the time later.

* Show draggable control bar in replay view

- fixes #949