Slightly better replay handling in minimized mode (#1008)
- refs #999
- We now set the player state to not playing so that the user can
select the play button in minimized mode in order to replay the
video. This is still not how it should be, not a complete fix, but
an improvement none the less. New behavior is: Click play button in
minimized view when the video is in "exited" state (replay view
showing in maximized view) the video is reset to the beginning of
playback, however the video is still paused at this point, so it
actually requires two clicks to get the desired re-playback. I plan
to revisit this in the next update (or within this one if I have
the time.